Some styling changes, MobileCarousel + CarouselImage moved into Carousel
