+++ components/organisms/Carousel/CarouselImage/CarouselImage.stories.tsx
