+++ components/organisms/Carousel/CarouselImage/CarouselImage.tsx
