+++ components/organisms/Carousel/CarouselImage/animation-variants.ts
