+++ components/organisms/Carousel/CarouselImage/index.tsx
