+++ components/organisms/Carousel/CarouselImage/styles.tsx
