+++ components/organisms/Carousel/MobileCarousel/MobileCarousel.stories.tsx
