+++ components/organisms/Carousel/MobileCarousel/styles.tsx
