+++ components/pages/HomePage/HomePage.tsx
@@ -10,7 +10,7 @@ import Carousel from "components/organisms/Carousel";
 import { Element } from "react-scroll";
 import { IDiscoverMovie } from "types/api/discover";
 import { IVideo } from "types/api/videos";
-import MobileCarousel from "components/organisms/MobileCarousel";
+import MobileCarousel from "components/organisms/Carousel/MobileCarousel";
 import { ReactElement } from "react";
 import ScrollDownBtn from "components/molecules/ScrollDownBtn";
 import { useMediaQuery } from "react-responsive";
